requirements-dev.txt
-r requirements.txt

pre-commit
pytest
pytest-approvaltests
tox

# Pin to match versions in precommit hooks
black==20.8b1
flake8==3.8.4
isort==5.2.1
